Remove deprecated schematize from SharedTreeView
