+++ experimental/dds/tree2/src/test/shared-tree/sharedTree.spec.ts
@@ -14,7 +14,6 @@ import {
 	singleTextCursor,
 	makeSchemaCodec,
 	jsonableTreeFromCursor,
-	on,
 	ContextuallyTypedNodeData,
 	Any,
 	TreeStatus,
@@ -30,7 +29,6 @@ import {
 	expectSchemaEqual,
 	initializeTestTree,
 	jsonSequenceRootSchema,
-	toJsonableTree,
 	validateTree,
 	validateTreeContent,
 	validateViewConsistency,
@@ -70,30 +68,89 @@ const schemaCodec = makeSchemaCodec({ jsonValidator: typeboxValidator });
 const fooKey: FieldKey = brand("foo");
 
 describe("SharedTree", () => {
-	// TODO: concurrent use of schematize should not double initialize. Should use constraints so second run conflicts.
-	it.skip("Concurrent Schematize", () => {
-		const provider = new TestTreeProviderLite(2);
-		const content: InitializeAndSchematizeConfiguration = {
-			schema: jsonSequenceRootSchema,
-			allowedSchemaModifications: AllowedUpdateType.None,
-			initialTree: [1],
-		};
-		const tree1 = provider.trees[0].schematizeView(content);
-		provider.trees[1].schematizeView(content);
-		provider.processMessages();
-
-		validateRootField(tree1, [1]);
-	});
-
-	it("reads only one node", () => {
-		// This is a regression test for a scenario in which a transaction would apply its delta twice,
-		// inserting two nodes instead of just one
-		const view = viewWithContent({ schema: jsonSequenceRootSchema, initialTree: [] });
-		runSynchronous(view, (t) => {
-			t.context.root.insertNodes(0, [5]);
+	describe("schematize", () => {
+		const factory = new SharedTreeFactory({
+			jsonValidator: typeboxValidator,
+			forest: ForestType.Reference,
 		});
 
-		assert.deepEqual(toJsonableTree(view), [{ type: leaf.number.name, value: 5 }]);
+		const builder = new SchemaBuilder({
+			scope: "test",
+			name: "Schematize Tree Tests",
+		});
+		const schema = builder.intoSchema(SchemaBuilder.optional(leaf.number));
+
+		const builderGeneralized = new SchemaBuilder({
+			scope: "test",
+			name: "Schematize Tree Tests Generalized",
+		});
+
+		const schemaGeneralized = builderGeneralized.intoSchema(SchemaBuilder.optional(Any));
+
+		// TODO: concurrent use of schematize should not double initialize. Should use constraints so second run conflicts.
+		it.skip("Concurrent Schematize", () => {
+			const provider = new TestTreeProviderLite(2);
+			const content: InitializeAndSchematizeConfiguration = {
+				schema: jsonSequenceRootSchema,
+				allowedSchemaModifications: AllowedUpdateType.None,
+				initialTree: [1],
+			};
+			const tree1 = provider.trees[0].schematizeView(content);
+			provider.trees[1].schematizeView(content);
+			provider.processMessages();
+
+			validateRootField(tree1, [1]);
+		});
+
+		it("initialize tree", () => {
+			const tree = factory.create(new MockFluidDataStoreRuntime(), "the tree");
+			assert.equal(tree.contentSnapshot().schema.rootFieldSchema, storedEmptyFieldSchema);
+
+			const view = tree.schematize({
+				allowedSchemaModifications: AllowedUpdateType.None,
+				initialTree: 10 as any,
+				schema,
+			});
+			assert.equal(view.content, 10);
+		});
+
+		it("noop upgrade", () => {
+			const tree = factory.create(new MockFluidDataStoreRuntime(), "the tree") as SharedTree;
+			tree.storedSchema.update(schema);
+
+			// No op upgrade with AllowedUpdateType.None does not error
+			const schematized = tree.schematizeView({
+				allowedSchemaModifications: AllowedUpdateType.None,
+				initialTree: 10,
+				schema,
+			});
+			// And does not add initial tree:
+			assert.equal(schematized.root, undefined);
+		});
+
+		it("incompatible upgrade errors", () => {
+			const tree = factory.create(new MockFluidDataStoreRuntime(), "the tree") as SharedTree;
+			tree.storedSchema.update(schemaGeneralized);
+			assert.throws(() => {
+				tree.schematize({
+					allowedSchemaModifications: AllowedUpdateType.None,
+					initialTree: 5,
+					schema,
+				});
+			});
+		});
+
+		it("upgrade schema", () => {
+			const tree = factory.create(new MockFluidDataStoreRuntime(), "the tree") as SharedTree;
+			tree.storedSchema.update(schema);
+			const schematized = tree.schematizeView({
+				allowedSchemaModifications: AllowedUpdateType.SchemaCompatible,
+				initialTree: 5,
+				schema: schemaGeneralized,
+			});
+			// Initial tree should not be applied
+			assert.equal(schematized.root, undefined);
+		});
 	});
 
 	it("editable-tree-2-end-to-end", () => {
@@ -905,92 +962,6 @@ describe("SharedTree", () => {
 		});
 		const schema = builder.intoSchema(builder.optional(Any));
 
-		it("triggers events for local and subtree changes", () => {
-			const view = viewWithContent({
-				schema,
-				initialTree: {
-					x: 24,
-				},
-			});
-			const rootNode = view.context.root.getNode(0);
-			const root = view.root as unknown as { x: number };
-			const log: string[] = [];
-			const unsubscribe = rootNode[on]("changing", () => log.push("change"));
-			const unsubscribeSubtree = rootNode[on]("subtreeChanging", () => {
-				log.push("subtree");
-			});
-			const unsubscribeAfter = view.events.on("afterBatch", () => log.push("after"));
-			log.push("editStart");
-			root.x = 5;
-			log.push("editStart");
-			root.x = 6;
-			log.push("unsubscribe");
-			unsubscribe();
-			unsubscribeSubtree();
-			unsubscribeAfter();
-			log.push("editStart");
-			root.x = 7;
-
-			assert.deepEqual(log, [
-				"editStart",
-				"subtree",
-				"subtree",
-				"change",
-				"after",
-				"editStart",
-				"subtree",
-				"subtree",
-				"change",
-				"after",
-				"unsubscribe",
-				"editStart",
-			]);
-		});
-
-		it("propagates path args for local and subtree changes", () => {
-			const view = viewWithContent({
-				schema,
-				initialTree: {
-					x: 24,
-				},
-			});
-			const rootNode = view.context.root.getNode(0);
-			const root = view.root as unknown as { x: number };
-			const log: string[] = [];
-			const unsubscribe = rootNode[on]("changing", (upPath) =>
-				log.push(`change-${String(upPath.parentField)}-${upPath.parentIndex}`),
-			);
-			const unsubscribeSubtree = rootNode[on]("subtreeChanging", (upPath) => {
-				log.push(`subtree-${String(upPath.parentField)}-${upPath.parentIndex}`);
-			});
-			const unsubscribeAfter = view.events.on("afterBatch", () => log.push("after"));
-			log.push("editStart");
-			root.x = 5;
-			log.push("editStart");
-			root.x = 6;
-			log.push("unsubscribe");
-			unsubscribe();
-			unsubscribeSubtree();
-			unsubscribeAfter();
-			log.push("editStart");
-			root.x = 7;
-
-			assert.deepEqual(log, [
-				"editStart",
-				"subtree-rootFieldKey-0",
-				"subtree-rootFieldKey-0",
-				"change-rootFieldKey-0",
-				"after",
-				"editStart",
-				"subtree-rootFieldKey-0",
-				"subtree-rootFieldKey-0",
-				"change-rootFieldKey-0",
-				"after",
-				"unsubscribe",
-				"editStart",
-			]);
-		});
-
 		it("triggers revertible events for local changes", () => {
 			const value = "42";
 			const provider = new TestTreeProviderLite(2);
@@ -1037,33 +1008,6 @@ describe("SharedTree", () => {
 			assert.equal(redoStack1.length, 0);
 			assert.equal(undoStack2.length, 1);
 			assert.equal(redoStack2.length, 0);
-
-			unsubscribe1();
-			unsubscribe2();
-		});
-
-		// TODO: unskip once forking revertibles is supported
-		it.skip("triggers a revertible event for a changes merged into the local branch", () => {
-			const tree1 = viewWithContent({
-				schema: jsonSequenceRootSchema,
-				initialTree: [],
-			});
-			const branch = tree1.fork();
-
-			const { undoStack: undoStack1, unsubscribe: unsubscribe1 } =
-				createTestUndoRedoStacks(tree1);
-			const { undoStack: undoStack2, unsubscribe: unsubscribe2 } =
-				createTestUndoRedoStacks(branch);
-
-			// Insert node
-			branch.setContent(["42"]);
-
-			assert.equal(undoStack1.length, 0);
-			assert.equal(undoStack2.length, 1);
-
-			tree1.merge(branch);
-			assert.equal(undoStack1.length, 1);
-			assert.equal(undoStack2.length, 1);
 
 			unsubscribe1();
 			unsubscribe2();
